[regexp] Split code point set builders and first match analysis out of regexp compiler
